Move around atool modules/packages
* Move most tiny packages into modules
* Move around some templates
* Move some functionality like auth out of atool
