+++ ocfweb/settings.py
@@ -26,13 +26,8 @@ INSTALLED_APPS = (
     'ocfweb',
     'ocfweb.about',
     'ocfweb.atool',
-    'ocfweb.atool.approve',
-    'ocfweb.atool.calnet',
-    'ocfweb.atool.chpass',
-    'ocfweb.atool.commands',
-    'ocfweb.atool.ocf',
-    'ocfweb.atool.vhost',
     'ocfweb.docs',
+    'ocfweb.login',
     'ocfweb.main',
     'ocfweb.middleware',
     'ocfweb.stats',
